breaks up reindexservice
