+++ common/src/main/scala/uk/ac/wellcome/utils/DynamoUtils.scala
@@ -34,3 +34,4 @@ trait DynamoUpdateWriteCapacityCapable {
     table.updateTable(newThroughput)
   }
 }
+
