+++ reindexer/src/main/scala/uk/ac/wellcome/platform/reindexer/models/ReindexAttempt.scala
@@ -1,0 +1,7 @@
+package uk.ac.wellcome.platform.reindexer.models
+
+import uk.ac.wellcome.models.{Reindex, Reindexable}
+
+case class ReindexAttempt(reindex: Reindex,
+                          successful: List[Reindexable[String]],
+                          attempt: Int)
