+++ reindexer/src/main/scala/uk/ac/wellcome/platform/reindexer/services/MiroReindexService.scala
@@ -1,0 +1,28 @@
+package uk.ac.wellcome.platform.reindexer.services
+
+import javax.inject.Inject
+
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
+import com.gu.scanamo.{Scanamo, Table}
+import com.twitter.inject.annotations.Flag
+import uk.ac.wellcome.models.MiroTransformable
+import uk.ac.wellcome.models.aws.DynamoConfig
+import uk.ac.wellcome.platform.reindexer.lib.ReindexService
+
+class MiroReindexService @Inject()(
+  reindexTrackerService: ReindexTrackerService,
+  dynamoDBClient: AmazonDynamoDB,
+  dynamoConfigs: Map[String, DynamoConfig],
+  @Flag("reindex.target.tableName") reindexTargetTableName: String)
+    extends ReindexService[MiroTransformable](reindexTrackerService,
+                                              dynamoDBClient,
+                                              dynamoConfigs,
+                                              reindexTargetTableName,
+                                              "miro") {
+
+  override val transformableTable: Table[MiroTransformable] =
+    Table[MiroTransformable](reindexTargetTableName)
+
+  override val scanamoQuery: ScanamoQuery =
+    Scanamo.queryIndex[MiroTransformable](dynamoDBClient) _
+}
