+++ reindexer/src/test/scala/uk/ac/wellcome/platform/reindexer/services/ReindexTrackerServiceTest.scala
@@ -1,0 +1,50 @@
+package uk.ac.wellcome.platform.reindexer.services
+
+import com.gu.scanamo.Scanamo
+import org.scalatest.{FunSpec, Matchers}
+import org.scalatest.concurrent.ScalaFutures
+import uk.ac.wellcome.models.Reindex
+import uk.ac.wellcome.models.aws.DynamoConfig
+import uk.ac.wellcome.test.utils.{DynamoDBLocal, ExtendedPatience}
+
+class ReindexTrackerServiceTest
+    extends FunSpec
+    with ScalaFutures
+    with Matchers
+    with DynamoDBLocal
+    with ExtendedPatience {
+
+  val dynamoConfigs = Map(
+    "reindex" -> DynamoConfig("applicationName",
+                              "streamArn",
+                              reindexTableName),
+    "calm" -> DynamoConfig("applicationName", "streamArn", calmDataTableName))
+
+  it(
+    "should return the index in need of reindexing"
+  ) {
+    val expectedReindex = Reindex("CalmData", 2, 1)
+    val reindexList = List(
+      expectedReindex,
+      Reindex("MiroData", 1, 1)
+    )
+
+    val expectedReindexList = List(
+      Reindex("foo", 2, 1)
+    )
+
+    reindexList.foreach(Scanamo.put(dynamoDbClient)(reindexTableName))
+
+    val reindexTrackerService = new ReindexTrackerService(
+      dynamoDbClient,
+      dynamoConfigs,
+      "CalmData"
+    )
+
+    val op = reindexTrackerService.getIndicesForReindex
+
+    whenReady(op) { (reindex: Option[Reindex]) =>
+      Some(expectedReindex) shouldBe reindex
+    }
+  }
+}
